cores/media_entertainment/predictive-CLTV-insights/pyproject.toml
[project]
name = "media_entertainment-predictive-cltv-insights"
version = "0.1.0"
description = "Predictive CLTV Insights"
industry = "Media Entertainment"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "accelerate>=0.26.0",
    "bertopic>=0.16.0",
    "bert-score>=0.3.13",
    "boto3>=1.28.0",
    "cloudpickle>=3.0.0",
    "cython>=3.2.2",
    "datasets>=2.14.1",
    "deltalake>=0.14.0",
    "groq>=0.37.0",
    "hdbscan>=0.8.33",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "kedro>=0.19.0",
    "kedro-datasets>=3.0",
    "kedro-viz>=6.7.0",
    "lifetimes>=0.11.3",
    "lifelines>=0.28.0",
    "matplotlib>=3.8.2",
    "mlflow>=2.8.0",
    "kedro-mlflow>=0.12.0,<0.14.0",
    "numpy>=1.26.2",
    "pandas>=2.1.4",
    "polars>=0.20.0",
    "plotly>=5.18.0",
    "prefect>=2.14.0",
    "prefect-aws>=0.4.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.4.0,<2.12.0",
    "pandera>=0.20.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "rich>=14.2.0",
    "scikit-learn>=1.3.2",
    "seaborn>=0.13.0",
    "sentencepiece>=0.1.99",
    "sentence-transformers>=3.5.0",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.29.0",
    "tqdm>=4.66.1",
    "transformers>4.36.0",
    "umap-learn>=0.5.5",
    "watchdog>=6.0.0",
    "xgboost>=2.0.0",
    "feast>=0.38.0",
]

[dependency-groups]
dev = [
    "black==23.9.0",
    "flake8==6.1.0",
    "mypy==1.5.0",
    "pre-commit==3.4.0",
    "pytest==8.3.4",
    "pytest-cov>=6.0.0",
]

[tool.kedro]
package_name = "ai_core"
project_name = "Predictive CLTV Insights"
kedro_init_version = "0.19.0"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/ai_core -ra"""
pythonpath = ["src"]

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.kedro_telemetry]
project_id = "a5647727e78642929a76c0a12a6ebe21"

[project.optional-dependencies]
xpu = [
    # Windows: Full XPU support with native PyTorch XPU backend
    "torch>=2.6.0; sys_platform == 'win32'",
    "torchvision>=0.21.0; sys_platform == 'win32'",
    "torchaudio>=2.6.0; sys_platform == 'win32'",
    "scikit-learn-intelex>=2025.10.0; sys_platform != 'darwin'",
    # Linux: XPU support requires manual installation due to missing pytorch-triton-xpu
    # See README.md for Linux XPU setup instructions (Intel compute runtime + pip install)
    # This extra is provided for cross-platform compatibility but won't install on Linux during resolution
]
cuda = ["torch>=2.6.0", "torchvision>=0.21.0", "torchaudio>=2.6.0"]
cpu = ["torch>=2.6.0", "torchvision>=0.21.0", "torchaudio>=2.6.0"]
mps = ["torch>=2.6.0", "torchvision>=0.21.0", "torchaudio>=2.6.0"]

[tool.uv]
index-strategy = "unsafe-best-match"
override-dependencies = ["lxml>=5.3.1; sys_platform == 'darwin'"]
environments = [
    "sys_platform == 'darwin'",
    "sys_platform == 'win32'",
    "sys_platform == 'linux'",
]
conflicts = [
    [
        { extra = "xpu" },
        { extra = "cuda" },
        { extra = "cpu" },
        { extra = "mps" },
    ],
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-xpu"
url = "https://download.pytorch.org/whl/xpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-xpu", extra = "xpu" },
    { index = "pytorch-cuda", extra = "cuda" },
    { index = "pytorch-cpu", extra = "cpu" },
]
torchvision = [
    { index = "pytorch-xpu", extra = "xpu" },
    { index = "pytorch-cuda", extra = "cuda" },
    { index = "pytorch-cpu", extra = "cpu" },
]
torchaudio = [
    { index = "pytorch-xpu", extra = "xpu" },
    { index = "pytorch-cuda", extra = "cuda" },
    { index = "pytorch-cpu", extra = "cpu" },
]
pytorch-triton-xpu = { index = "pytorch-xpu", marker = "platform_machine == 'x86_64' and sys_platform == 'linux'" }
triton-xpu = { index = "pytorch-xpu" }
intel-extension-for-pytorch = { index = "pytorch-xpu" }
intel-sycl-rt = { index = "pytorch-xpu" }
dpcpp-cpp-rt = { index = "pytorch-xpu" }
intel-cmplr-lib-rt = { index = "pytorch-xpu" }
intel-cmplr-lic-rt = { index = "pytorch-xpu" }
intel-cmplr-lib-ur = { index = "pytorch-xpu" }
